Seguimiento de linea con PID SIN CALIBRACION DINAMICA FUNCIONANDO PERO PID MAS RAPIDO Y SERVO ARREGLADO

+++ Core/Inc/Servo.h
@@ -10,28 +10,29 @@ typedef struct {
     servo_write_pwm_cb_t set_pwm;
 
     // Calibración del SG90 (En microsegundos)
-    uint16_t min_pulse_us; // Típicamente 500
-    uint16_t max_pulse_us; // Típicamente 2500
+    uint16_t min_pulse_us;
+    uint16_t max_pulse_us;
 
-    // Variables de estado para movimiento suave y no bloqueante
-    float current_angle;
-    float target_angle;
-    float speed_deg_per_tick;
+    // Variables de estado en PUNTO FIJO (Ángulo * 100)
+    // Ejemplo: 90 grados = 9000, 180 grados = 18000
+    int32_t current_angle;
+    int32_t target_angle;
+    int32_t speed_deg_per_tick; // Velocidad en centésimas de grado por tick
 } sServoHandle;
 
-// Inicializa el servo y lo lleva al centro (90 grados)
+// Inicializa el servo y lo lleva al centro (9000 = 90.00 grados)
 void Servo_Init(sServoHandle *dev, servo_write_pwm_cb_t hardware_writer);
 
-// Define los límites reales de tu servo (algunos chinos varían entre 600 y 2400)
+// Define los límites reales de tu servo
 void Servo_Calibrate(sServoHandle *dev, uint16_t min_us, uint16_t max_us);
 
-// Mueve instantáneamente el servo a un ángulo (0 a 180)
-void Servo_SetAngle(sServoHandle *dev, float angle);
+// Mueve instantáneamente el servo a un ángulo en punto fijo (0 a 18000)
+void Servo_SetAngle(sServoHandle *dev, int32_t angle);
 
-// Establece un ángulo objetivo y a qué velocidad debe viajar
-void Servo_MoveToSmooth(sServoHandle *dev, float target_angle, float speed);
+// Establece un ángulo objetivo y a qué velocidad debe viajar (valores * 100)
+void Servo_MoveToSmooth(sServoHandle *dev, int32_t target_angle, int32_t speed);
 
-// Tarea que se debe llamar en el while(1) periódicamente para los movimientos suaves
+// Tarea periódica para los movimientos suaves (llamar en el while(1))
 void Servo_Task(sServoHandle *dev);
 
 #endif /* SERVO_H_ */
